Rename package to configprovider to avoid collision with core configsource
