+++ internal/configprovider/helpers.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package configsource
+package configprovider
 
 // WatcherNotSupported is the a watcher function that always returns ErrWatcherNotSupported.
 func WatcherNotSupported() error {
